mongo-client: Fix a signedness warning.
In mongo_packet_recv(), cast the return value of recv to guint32, to
silence a warning.

It's safe to do that, because if recv fails and returns -1, that's not
going to be equal to size.

+++ src/mongo-client.c
@@ -251,7 +251,7 @@ mongo_packet_recv (mongo_connection *conn)
 
   size = h.length - sizeof (mongo_packet_header);
   data = g_try_new0 (guint8, size);
-  if (recv (conn->fd, data, size, MSG_NOSIGNAL) != size)
+  if ((guint32)recv (conn->fd, data, size, MSG_NOSIGNAL) != size)
     {
       int e = errno;
